pyproject.toml
[project]
name = "pytorchlightning-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lightning>=2.5.5",
    "matplotlib>=3.10.7",
    "natsort>=8.4.0",
    "onnxruntime-gpu>=1.23.2",
    "scipy>=1.15.3",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "toml>=0.10.2",
    "torchsummary>=1.5.1",
    "torchsummaryx>=1.3.0",
    "torchvision>=0.24.0",
]
